rocket_launch_bot/requirements.txt
python-telegram-bot==20.7
httpx>=0.24.0
Pillow>=10.0.0
python-dotenv>=0.19.0